Add script to generate `library/gradle` contents
This is a little bit complex, but so is the matrix of supported variations.  It could be simplified somewhat by maintaining all supported versions in a single branch so they could all be checked out concurrently, but it manages OK without that.

The "generate a list of tags" logic is a little bit complex because I tried to match the existing tag ordering / existence, only filling in holes where there was pre-existing precedent (such as "X-jdk-graal" existing but not "X.Y-jdk-graal").

There are a few TODOs in the code; most notably, there are a few directories on some of the upstream branches that should be deleted so the script doesn't pick them up (and then the related code can be removed or refactored to error instead).  Also, I did not implement the "set intersection" necessary to get `Architectures:` 100% correct for "jdk-lts-and-current" variants because they currently all match -- if that ever isn't the case, it shouldn't be too hard to implement (I usually do it in `jq` where it's really easy, if a bit esoteric).